requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.3.0
bs4==0.0.1
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.5
django-gsheets==0.0.10
djangorestframework==3.13.1
flower==1.0.0
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
html5lib==1.1
httplib2==0.20.4
humanize==4.2.1
idna==3.3
kombu==5.2.4
lxml==4.9.0
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.21.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
tornado==6.1
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
